Split france/__init__.py into multiple files, optimize DBPedia for bulk query, process EPCI data
