+++ geozones/dbpedia.py
@@ -1,129 +1,84 @@
 import re
 import json
-from string import Template
 
 import requests
 
-from .tools import warning
+from .tools import error
 
+
+RDF_PREFIXES = {
+    'o:': 'http://dbpedia.org/ontology/',
+    'fr:': 'http://fr.dbpedia.org/property/',
+}
+
+SPARQL_PREFIXES = '\n'.join(map(
+    lambda kv: 'PREFIX {0[0]} <{0[1]}>'.format(kv),
+    RDF_PREFIXES.items()
+))
+
+# Sparql query for bulk properties retrieval
+# given a list of DBPedia URIs
+SPARQL_BY_URI = '''{prefixes}
+SELECT ?uri ?property ?value WHERE {{
+    VALUES ?uri {{ {uris} }}
+    VALUES ?property {{ {properties} }}
+    ?uri ?property ?value.
+}}
+'''
 
 RE_WIKIPEDIA = re.compile(
-    r'https?://(?P<namespace>\w+)?\.?wikipedia\.org/wiki/(?P<resource>.+)$')
+    r'https?://(?P<namespace>\w+)?\.?wikipedia\.org/wiki/(?P<path>.+)$')
+
 SPARQL_SERVER = 'http://dbpedia.inria.fr/sparql'
-# We want population and/or area from French DBPedia or
-# their international counterparts as fallbacks.
-SPARQL_POPULATION_TEMPLATE = Template('''SELECT ?population ?area WHERE {
-    {<$resource_url> <http://fr.dbpedia.org/property/population>|
-                     <http://dbpedia.org/ontology/populationTotal> ?population}
-UNION
-    {<$resource_url> <http://fr.dbpedia.org/property/superficie>|
-                     <http://dbpedia.org/ontology/area> ?area}
-}''')
-
-# We want flag and/or blason from French DBPedia or
-# their international counterparts as fallbacks.
-SPARQL_IMAGE_TEMPLATE = Template('''SELECT ?flag ?blazon WHERE {
-    {<$resource_url> <http://dbpedia.org/ontology/flag> ?flag}
-UNION
-    {<$resource_url> <http://dbpedia.org/ontology/blazon> ?blazon}
-}''')
 
 
-class DBPedia(object):
+def extract_uri(uri):
+    '''Extract a DBPedia URI from a Wikipedia identifier or URL'''
+    if not uri:
+        return
+    uri = uri.strip().replace(' ', '_')
+    # Special wrong case: `fr:fr:Communauté_de_communes_d'Altkirch`
+    if uri.startswith('fr:fr:'):
+        namespace, _, path = uri.split(':')
+    elif ':' in uri and not uri.startswith('http'):
+        namespace, path = uri.split(':')
+    elif RE_WIKIPEDIA.match(uri):
+        m = RE_WIKIPEDIA.match(uri)
+        namespace, path = m.group('namespace', 'path')
+    else:
+        path = uri
+        namespace = None
 
-    def __init__(self, resource):
-        resource = resource.strip('').replace(' ', '_')
-        # Special wrong case: `fr:fr:Communauté_de_communes_d'Altkirch`
-        if resource.startswith('fr:fr:'):
-            namespace, _, self.resource = resource.split(':')
-        elif ':' in resource and not resource.startswith('http'):
-            namespace, self.resource = resource.split(':')
-        elif RE_WIKIPEDIA.match(resource):
-            m = RE_WIKIPEDIA.match(resource)
-            namespace, self.resource = m.group('namespace', 'resource')
-        else:
-            self.resource = resource
-            namespace = None
+    if namespace:
+        base_url = 'http://{0}.dbpedia.org'.format(namespace)
+    else:
+        base_url = 'http://dbpedia.org'
+    return '{base_url}/resource/{path}'.format(base_url=base_url, path=path)
 
-        if namespace:
-            self.base_url = 'http://{0}.dbpedia.org'.format(namespace)
-        else:
-            self.base_url = 'http://dbpedia.org'
-        self.resource_url = '{base_url}/resource/{resource}'.format(
-            base_url=self.base_url, resource=self.resource)
 
-    def fetch_population_or_area(self):
-        sparql_query = SPARQL_POPULATION_TEMPLATE.substitute(
-            resource_url=self.resource_url)
-        parameters = {
-            'default-graph-uri': 'http://fr.dbpedia.org',
-            'query': sparql_query,
-            'format': 'json'
-        }
-        result = {}
-        try:
-            response = requests.get(SPARQL_SERVER, params=parameters)
-        except requests.exceptions.ReadTimeout:
-            warning('Timeout:', SPARQL_SERVER, parameters)
-            return result
-        try:
-            data = response.json()
-        except json.decoder.JSONDecodeError:
-            warning('JSON Error: {0} {1} {2}',
-                    SPARQL_SERVER, parameters, response.text)
-            return result
-        try:
-            results = data['results']['bindings'][0]
-        except IndexError:
-            return result
-        if 'population' in results:
-            try:
-                result['population'] = int(results['population']['value'])
-            except ValueError:
-                pass
-        if 'area' in results:
-            try:
-                result['area'] = int(round(float(results['area']['value'])))
-            except ValueError:
-                pass
-        return result
-
-    def fetch_flag_or_blazon(self):
-        sparql_query = SPARQL_IMAGE_TEMPLATE.substitute(
-            resource_url=self.resource_url)
-        parameters = {
-            'default-graph-uri': 'http://fr.dbpedia.org',
-            'query': sparql_query,
-            'format': 'json'
-        }
-        flag_or_blazon = {}
-        try:
-            response = requests.get(SPARQL_SERVER, params=parameters)
-        except requests.exceptions.ReadTimeout:
-            warning('Timeout:', SPARQL_SERVER, parameters)
-            return flag_or_blazon
-        try:
-            data = response.json()
-        except json.decoder.JSONDecodeError:
-            warning('JSON Error: {0} {1} {2}',
-                    SPARQL_SERVER, parameters, response.text)
-            return flag_or_blazon
-        try:
-            results = data['results']['bindings']
-        except IndexError:
-            return flag_or_blazon
-        if not results:
-            return flag_or_blazon
-        if 'flag' in results[0]:
-            flag_name = results[0]['flag']['value'].replace(' ', '_')
-            flag_or_blazon['flag'] = flag_name
-            if len(results) > 1:
-                blazon_name = results[1]['blazon']['value'].replace(' ', '_')
-                flag_or_blazon['blazon'] = blazon_name
-        if 'blazon' in results[0]:
-            blazon_name = results[0]['blazon']['value'].replace(' ', '_')
-            flag_or_blazon['blazon'] = blazon_name
-            if len(results) > 1:
-                flag_name = results[1]['flag']['value'].replace(' ', '_')
-                flag_or_blazon['flag'] = flag_name
-        return flag_or_blazon
+def execute_sparql_query(query, graph='http://fr.dbpedia.org'):
+    '''
+    Execute a SPARQL query and returns a list of n-uplets.
+    '''
+    headers = {
+        'User-Agent': 'geozones/1.0 (http://github.com/etalab/geozones)',
+    }
+    parameters = {
+        'default-graph-uri': graph,
+        'query': query,
+        'format': 'json'
+    }
+    try:
+        response = requests.post(SPARQL_SERVER,
+                                 data=parameters,
+                                 headers=headers)
+    except requests.exceptions.ReadTimeout:
+        error('Timeout:', SPARQL_SERVER, parameters)
+        return []
+    try:
+        data = response.json()
+    except json.decoder.JSONDecodeError:
+        error('JSON Error: {0} {1} {2}',
+              SPARQL_SERVER, parameters, response.text)
+        return []
+    return data['results']['bindings']
